:recycle: [#4606] Split into separate components for error boundary management
The retrieval of the available catalogues needs to be lifted up, so the
failure potential lifts up and we need to split into more components
so that we can wrap only part of the components in the error boundary,
otherwise users would not be able to select a different api group.
