+++ src/openforms/js/components/admin/form_design/registrations/zgw/fields/CatalogueSelect.js
@@ -1,42 +1,11 @@
 import {useFormikContext} from 'formik';
+import PropTypes from 'prop-types';
 import {FormattedMessage} from 'react-intl';
-import {useAsync} from 'react-use';
 
-import {
-  CatalogueSelect as GenericCatalogueSelect,
-  groupAndSortCatalogueOptions,
-} from 'components/admin/forms/zgw';
-import {get} from 'utils/fetch';
+import {CatalogueSelect as GenericCatalogueSelect} from 'components/admin/forms/zgw';
 
-// Data fetching
-
-const CATALOGUES_ENDPOINT = '/api/v2/registration/plugins/zgw-api/catalogues';
-
-const getCatalogues = async apiGroupID => {
-  const response = await get(CATALOGUES_ENDPOINT, {zgw_api_group: apiGroupID});
-  if (!response.ok) {
-    throw new Error('Loading available catalogues failed');
-  }
-  return groupAndSortCatalogueOptions(response.data);
-};
-
-// Components
-
-const CatalogueSelect = () => {
-  const {values} = useFormikContext();
-  const {zgwApiGroup = null} = values;
-
-  // fetch available catalogues
-  const {
-    loading: loadingCatalogues,
-    value: catalogueOptionGroups = [],
-    error: cataloguesError,
-  } = useAsync(async () => {
-    if (!zgwApiGroup) return [];
-    return await getCatalogues(zgwApiGroup);
-  }, [zgwApiGroup]);
-  if (cataloguesError) throw cataloguesError;
-
+const CatalogueSelect = ({loading, optionGroups}) => {
+  const {values: zgwApiGroup = null} = useFormikContext();
   // TODO: make required when case type can be selected in a dropdown AND there is no
   // legacy case type or document type URL specified. Probably best to do this as
   // backend validation so that new registration backends must select a catalogue.
@@ -50,12 +19,26 @@ const CatalogueSelect = () => {
         />
       }
       isDisabled={!zgwApiGroup}
-      loading={loadingCatalogues}
-      optionGroups={catalogueOptionGroups}
+      loading={loading}
+      optionGroups={optionGroups}
     />
   );
 };
 
-CatalogueSelect.propTypes = {};
+CatalogueSelect.propTypes = {
+  loading: PropTypes.bool.isRequired,
+  optionGroups: PropTypes.arrayOf(
+    PropTypes.shape({
+      label: PropTypes.string.isRequired,
+      options: PropTypes.arrayOf(
+        PropTypes.shape({
+          rsin: PropTypes.string.isRequired,
+          domain: PropTypes.string.isRequired,
+          label: PropTypes.string.isRequired,
+        })
+      ).isRequired,
+    })
+  ),
+};
 
 export default CatalogueSelect;
